Many many changes
So I'm still learning Laravel.  Messing around with it.  Trying to
figure out it's templating system and all of that.

This version is best viewed on an iphone for now.  works on the browser
too… just shrink your screen down.

A ton of new styles were introduced. Still trying to keep it really
light.